rename postgres reconcile file (#21364)
